requirements.txt
torch
pytorch-lightning
albumentations
torchvision
click
pathlib
numpy
matplotlib
PyYAML
scikit-learn
hydra-core
omegaconf
wandb